rename mdapi helper file
