+++ client/components/mma/cancel/cancellationSaves/CancellationSaves.stories.tsx
@@ -2,8 +2,8 @@ import type { ComponentMeta, ComponentStory } from '@storybook/react';
 import { rest } from 'msw';
 import { ReactRouterDecorator } from '../../../../../.storybook/ReactRouterDecorator';
 import { PRODUCT_TYPES } from '../../../../../shared/productTypes';
+import { toMembersDataApiResponse } from '../../../../fixtures/mdapiResponse';
 import { membershipSupporter } from '../../../../fixtures/productBuilder/testProducts';
-import { toMembersDataApiResponse } from '../../../../fixtures/productDetail';
 import { CancellationContainer } from '../CancellationContainer';
 import { ConfirmMembershipCancellation } from './ConfirmMembershipCancellation';
 import { ContinueMembershipConfirmation } from './ContinueMembershipConfirmation';
