+++ client/components/mma/delivery/address/DeliveryAddress.stories.tsx
@@ -2,8 +2,8 @@ import type { ComponentMeta, ComponentStory } from '@storybook/react';
 import { rest } from 'msw';
 import { ReactRouterDecorator } from '../../../../../.storybook/ReactRouterDecorator';
 import { PRODUCT_TYPES } from '../../../../../shared/productTypes';
+import { toMembersDataApiResponse } from '../../../../fixtures/mdapiResponse';
 import { guardianWeeklyPaidByCard } from '../../../../fixtures/productBuilder/testProducts';
-import { toMembersDataApiResponse } from '../../../../fixtures/productDetail';
 import { DeliveryAddressChangeContainer } from './DeliveryAddressChangeContainer';
 import { DeliveryAddressConfirmation } from './DeliveryAddressConfirmation';
 import { DeliveryAddressUpdate } from './DeliveryAddressForm';
