+++ client/components/mma/holiday/Holiday.stories.tsx
@@ -3,8 +3,8 @@ import { rest } from 'msw';
 import { ReactRouterDecorator } from '../../../../.storybook/ReactRouterDecorator';
 import { PRODUCT_TYPES } from '../../../../shared/productTypes';
 import { existingHolidays } from '../../../fixtures/holidays';
+import { toMembersDataApiResponse } from '../../../fixtures/mdapiResponse';
 import { guardianWeeklyPaidByCard } from '../../../fixtures/productBuilder/testProducts';
-import { toMembersDataApiResponse } from '../../../fixtures/productDetail';
 import { HolidayDateChooser } from './HolidayDateChooser';
 import { HolidaysOverview } from './HolidaysOverview';
 import { HolidayStopsContainer } from './HolidayStopsContainer';
