+++ client/fixtures/mdapiResponse.ts
@@ -1,13 +1,3 @@
-/*
- * Product details for signed in user
- * /api/me/mma
- *
- * Products are split out here, but will be returned as an array from the API.
- * eg. `[guardianWeeklyCard, digitalDD, newspaperVoucherPaypal]`
- *
- * DEPRECATED IN FAVOUR OF `productBuilder.ts`
- */
-
 import type {
 	MembersDataApiResponse,
 	ProductDetail,
