+++ cypress/integration/parallel-1/accountOverview.spec.ts
@@ -1,4 +1,4 @@
-import { toMembersDataApiResponse } from '../../../client/fixtures/productDetail';
+import { toMembersDataApiResponse } from '../../../client/fixtures/mdapiResponse';
 import { signInAndAcceptCookies } from '../../lib/signInAndAcceptCookies';
 import { singleContributionsAPIResponse } from '../../../client/fixtures/singleContribution';
 import { featureSwitches } from '../../../shared/featureSwitches';
