+++ cypress/integration/parallel-2/cancelSupporterPlus.spec.ts
@@ -1,5 +1,5 @@
 import { supporterPlus } from '../../../client/fixtures/productBuilder/testProducts';
-import { toMembersDataApiResponse } from '../../../client/fixtures/productDetail';
+import { toMembersDataApiResponse } from '../../../client/fixtures/mdapiResponse';
 import { signInAndAcceptCookies } from '../../lib/signInAndAcceptCookies';
 
 describe('Cancel Supporter Plus', () => {
