+++ cypress/integration/parallel-2/productSwitch.spec.ts
@@ -1,4 +1,4 @@
-import { toMembersDataApiResponse } from '../../../client/fixtures/productDetail';
+import { toMembersDataApiResponse } from '../../../client/fixtures/mdapiResponse';
 import { signInAndAcceptCookies } from '../../lib/signInAndAcceptCookies';
 import {
 	productMovePreviewResponse,
