+++ cypress/integration/parallel-2/updateContributionAmount.spec.ts
@@ -1,5 +1,5 @@
 import { contributionPaidByCard } from '../../../client/fixtures/productBuilder/testProducts';
-import { toMembersDataApiResponse } from '../../../client/fixtures/productDetail';
+import { toMembersDataApiResponse } from '../../../client/fixtures/mdapiResponse';
 import { signInAndAcceptCookies } from '../../lib/signInAndAcceptCookies';
 
 describe('Update contribution amount', () => {
