+++ cypress/integration/parallel-3/holidayStops.spec.ts
@@ -1,4 +1,4 @@
-import { toMembersDataApiResponse } from '../../../client/fixtures/productDetail';
+import { toMembersDataApiResponse } from '../../../client/fixtures/mdapiResponse';
 import {
 	potentialDeliveries,
 	noPotentialDeliveries,
