+++ cypress/integration/parallel-4/deliveryRecords.spec.ts
@@ -1,4 +1,4 @@
-import { toMembersDataApiResponse } from '../../../client/fixtures/productDetail';
+import { toMembersDataApiResponse } from '../../../client/fixtures/mdapiResponse';
 import {
 	deliveryRecordsWithNoDeliveries,
 	deliveryRecordsWithDelivery,
