+++ cypress/integration/parallel-5/membership.spec.ts
@@ -1,5 +1,5 @@
 import { membershipSupporter } from '../../../client/fixtures/productBuilder/testProducts';
-import { toMembersDataApiResponse } from '../../../client/fixtures/productDetail';
+import { toMembersDataApiResponse } from '../../../client/fixtures/mdapiResponse';
 import { signInAndAcceptCookies } from '../../lib/signInAndAcceptCookies';
 
 describe('membership test', () => {
