+++ cypress/lib/signInAndAcceptCookies.ts
@@ -1,5 +1,5 @@
 import { guardianWeeklyPaidByCard } from '../../client/fixtures/productBuilder/testProducts';
-import { toMembersDataApiResponse } from '../../client/fixtures/productDetail';
+import { toMembersDataApiResponse } from '../../client/fixtures/mdapiResponse';
 
 export const signInAndAcceptCookies = () => {
 	cy.session('auth', () => {
